pyproject.toml
[project]
name = "comic-book-reading-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
